# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boutique Stock Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  description?: string | null;
}

export interface Product {
  id: number;
  name: string;
  categoryId?: number | null;
  categoryName?: string | null;
  quantity: number;
  unitCostPrice?: number | null;
  unitSalePrice?: number | null;
  lowStockThreshold: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  categoryId?: number | null;
  quantity?: number;
  unitCostPrice?: number | null;
  unitSalePrice?: number | null;
  lowStockThreshold?: number;
  notes?: string | null;
}

export interface UpdateProductBody {
  name?: string;
  categoryId?: number | null;
  quantity?: number;
  unitCostPrice?: number | null;
  unitSalePrice?: number | null;
  lowStockThreshold?: number;
  notes?: string | null;
}

export interface AdjustStockBody {
  /** Positive to add, negative to remove */
  delta: number;
  reason?: string | null;
}

export interface Sale {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalAmount: number;
  notes?: string | null;
  createdAt: string;
}

export interface CreateSaleBody {
  productId: number;
  quantity: number;
  unitPrice: number;
  notes?: string | null;
}

export type StockMovementType =
  (typeof StockMovementType)[keyof typeof StockMovementType];

export const StockMovementType = {
  entry: "entry",
  exit: "exit",
  adjustment: "adjustment",
  sale: "sale",
} as const;

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  type: StockMovementType;
  delta: number;
  quantityBefore: number;
  quantityAfter: number;
  reason?: string | null;
  createdAt: string;
}

export interface CreateStockMovementBody {
  productId: number;
  /** Positive for entries, negative for exits */
  delta: number;
  reason?: string | null;
}

export interface DashboardSummary {
  totalProducts: number;
  totalUnits: number;
  totalStockValue: number;
  totalSalesValue: number;
  totalSalesCount: number;
  lowStockCount: number;
  outOfStockCount: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalSold: number;
  totalRevenue: number;
  categoryName?: string | null;
}

export interface CategoryBreakdown {
  categoryId?: number | null;
  categoryName: string;
  productCount: number;
  totalUnits: number;
  totalValue: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  sale: "sale",
  stock_entry: "stock_entry",
  adjustment: "adjustment",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  productId: number;
  productName: string;
  delta: number;
  description: string;
  createdAt: string;
}

export type ListProductsParams = {
  search?: string;
  categoryId?: number;
  lowStock?: boolean;
};

export type ListSalesParams = {
  limit?: number;
  offset?: number;
};

export type ListStockMovementsParams = {
  productId?: number;
  limit?: number;
  offset?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};