# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boutique Stock Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: categories
    description: Product categories
  - name: sales
    description: Sales recording
  - name: stock-movements
    description: Stock entries and movements
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /products/{id}/adjust-stock:
    post:
      operationId: adjustProductStock
      tags: [products]
      summary: Adjust product stock quantity (+/-)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustStockBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale (auto-deducts stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  /sales/{id}:
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Cancel a sale (restores stock)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stock-movements:
    get:
      operationId: listStockMovements
      tags: [stock-movements]
      summary: List all stock movements (history)
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [stock-movements]
      summary: Record a stock entry (restock / new collection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockMovementBody"
      responses:
        "201":
          description: Created stock movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary (stock value, counts, alerts)
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top-selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top products by sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"

  /dashboard/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [dashboard]
      summary: Stock value and quantity breakdown by category
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent stock movements and sales
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt

    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required:
        - name

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        quantity:
          type: integer
        unitCostPrice:
          type: number
          nullable: true
        unitSalePrice:
          type: number
          nullable: true
        lowStockThreshold:
          type: integer
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - quantity
        - lowStockThreshold
        - createdAt
        - updatedAt

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
          nullable: true
        quantity:
          type: integer
          default: 0
        unitCostPrice:
          type: number
          nullable: true
        unitSalePrice:
          type: number
          nullable: true
        lowStockThreshold:
          type: integer
          default: 2
        notes:
          type: string
          nullable: true
      required:
        - name

    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
          nullable: true
        quantity:
          type: integer
        unitCostPrice:
          type: number
          nullable: true
        unitSalePrice:
          type: number
          nullable: true
        lowStockThreshold:
          type: integer
        notes:
          type: string
          nullable: true

    AdjustStockBody:
      type: object
      properties:
        delta:
          type: integer
          description: Positive to add, negative to remove
        reason:
          type: string
          nullable: true
      required:
        - delta

    Sale:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
        - totalAmount
        - createdAt

    CreateSaleBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - productId
        - quantity
        - unitPrice

    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
          enum: [entry, exit, adjustment, sale]
        delta:
          type: integer
        quantityBefore:
          type: integer
        quantityAfter:
          type: integer
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - productName
        - type
        - delta
        - quantityBefore
        - quantityAfter
        - createdAt

    CreateStockMovementBody:
      type: object
      properties:
        productId:
          type: integer
        delta:
          type: integer
          description: Positive for entries, negative for exits
        reason:
          type: string
          nullable: true
      required:
        - productId
        - delta

    DashboardSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        totalUnits:
          type: integer
        totalStockValue:
          type: number
        totalSalesValue:
          type: number
        totalSalesCount:
          type: integer
        lowStockCount:
          type: integer
        outOfStockCount:
          type: integer
      required:
        - totalProducts
        - totalUnits
        - totalStockValue
        - totalSalesValue
        - totalSalesCount
        - lowStockCount
        - outOfStockCount

    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
        categoryName:
          type: string
          nullable: true
      required:
        - productId
        - productName
        - totalSold
        - totalRevenue

    CategoryBreakdown:
      type: object
      properties:
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
        productCount:
          type: integer
        totalUnits:
          type: integer
        totalValue:
          type: number
      required:
        - categoryName
        - productCount
        - totalUnits
        - totalValue

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [sale, stock_entry, adjustment]
        productId:
          type: integer
        productName:
          type: string
        delta:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - productId
        - productName
        - delta
        - description
        - createdAt